Support NWN:EE module folders (closes #8)
NWN:EE added module folders. These are subdirectories of the modules
folder that contain unpacked modules. They can be loaded in the toolset
just like a packed module.

This update allows nasher to make use of this feature. It can be enabled
using the new --useModuleFolder flag. The default value for this flag is
true during installation; during unpacking, it is true if the user has
not explicity chosen a file to unpack. If you are not using NWN:EE, you
should set this flag to false using nasher config:

    $ nasher config useModuleFolder false

+++ src/nasher/unpack.nim
@@ -119,18 +119,33 @@ proc unpack*(opts: Options, pkg: PackageRef) =
     fatal(fmt"Cannot unpack {file}: file does not exist")
 
   let
-    tmpDir = ".nasher" / "tmp"
-    fileName = file.extractFilename
+    (_, name, ext) = file.splitFile
+    shortFile = file.extractFilename
     erfUtil = opts.get("erfUtil")
     erfFlags = opts.get("erfFlags")
+    useFolder =
+      ext == ".mod" and
+      opts.get("useModuleFolder", not opts.hasKey("file"))
 
-  display("Extracting", fmt"{fileName} to {dir} using target {target.name}")
+  var
+    tmpDir = ".nasher" / "tmp"
+
+  display("Extracting", fmt"{shortFile} to {dir} using target {target.name}")
   setCurrentDir(dir)
-  removeDir(tmpDir)
-  createDir(tmpDir)
 
-  withDir(tmpDir):
-    extractErf(file, erfUtil, erfFlags)
+  if useFolder:
+    tmpDir = installDir / "modules" / name
+
+    info("Using", "module folder at " & tmpDir)
+    if not existsDir(tmpDir):
+      createDir(tmpDir)
+      withDir(tmpDir):
+        extractErf(file, erfUtil, erfFlags)
+  else:
+    removeDir(tmpDir)
+    createDir(tmpDir)
+    withDir(tmpDir):
+      extractErf(file, erfUtil, erfFlags)
 
   var
     manifest = parseManifest(target.name)
@@ -140,7 +155,6 @@ proc unpack*(opts: Options, pkg: PackageRef) =
     sourceFiles = getSourceFiles(target.includes, target.excludes)
     srcMap = genSrcMap(sourceFiles)
     packTime = file.getLastModificationTime
-    shortFile = file.extractFilename
     removeDeleted = opts.get("removeDeleted", false)
     askRemove = not opts.hasKey("removeDeleted")
 
@@ -160,9 +174,10 @@ proc unpack*(opts: Options, pkg: PackageRef) =
     if ext in GffExtensions:
       sourceName.add("." & gffFormat)
 
-    if sourceName notin sourceFiles and existsFile(tmpDir/fileName):
+    if sourceName notin sourceFiles and existsFile(tmpDir / fileName):
       if not askIf(fmt"{fileName} not found in source directory. Should it be re-added?"):
-        removeFile(tmpDir/fileName)
+        if not useFolder:
+          removeFile(tmpDir / fileName)
 
       # Delete from manifest whether yes or no, because if no we want it found
       # and re-added via changedFiles
